Move EncodeCSV to S3 plugin
