+++ plugins/s3/csv_test.go
@@ -1,11 +1,16 @@
 package s3
 
 import (
+	"bytes"
+	"encoding/csv"
 	"fmt"
 	"io"
+	"io/ioutil"
 	"strings"
+	"testing"
 	"time"
 
+	"github.com/stretchr/testify/assert"
 	"github.com/stripe/veneur/samplers"
 )
 
@@ -17,7 +22,7 @@ type CSVTestCase struct {
 
 func CSVTestCases() []CSVTestCase {
 
-	partition := time.Now().Format("20060102")
+	partition := time.Now().UTC().Format("20060102")
 
 	return []CSVTestCase{
 		{
@@ -70,4 +75,52 @@ func CSVTestCases() []CSVTestCase {
 			Row: strings.NewReader(fmt.Sprintf("a.b.c.max\t\"{foo:b\tar,baz:quz}\"\trate\tlocalhost\ttestbox-c3eac9\teniac\t10\t2016-10-10 05:04:18\t100\t%s\n", partition)),
 		},
 	}
+}
+
+func TestEncodeCSV(t *testing.T) {
+	testCases := CSVTestCases()
+
+	for _, tc := range testCases {
+		t.Run(tc.Name, func(t *testing.T) {
+
+			b := &bytes.Buffer{}
+
+			w := csv.NewWriter(b)
+			w.Comma = '\t'
+
+			tm := time.Now()
+			err := EncodeDDMetricCSV(tc.DDMetric, w, &tm, "testbox-c3eac9")
+			assert.NoError(t, err)
+
+			// We need to flush or there won't actually be any data there
+			w.Flush()
+			assert.NoError(t, err)
+
+			assertReadersEqual(t, tc.Row, b)
+		})
+	}
+}
+
+// Helper function for determining that two readers are equal
+func assertReadersEqual(t *testing.T, expected io.Reader, actual io.Reader) {
+
+	// If we can seek, ensure that we're starting at the beginning
+	for _, reader := range []io.Reader{expected, actual} {
+		if readerSeeker, ok := reader.(io.ReadSeeker); ok {
+			readerSeeker.Seek(0, io.SeekStart)
+		}
+	}
+
+	// do the lazy thing for now
+	bts, err := ioutil.ReadAll(expected)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	bts2, err := ioutil.ReadAll(actual)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	assert.Equal(t, string(bts), string(bts2))
 }
